deprecated {pillar}	R Documentation

Deprecated functions

Description

is_vector_s3() returns TRUE if an object represents an S3 vector. The default implementation forwards to vctrs::vec_is().

Usage

is_vector_s3(x)

Lifecycle

is_vector_s3() is soft-deprecated, because vctrs::vec_is() provides the same functionality based on a robust framework. The generic will remain available for a while but is no longer used. Users and implementers are encouraged to ensure that vctrs::vec_is() is TRUE for their class.

[Package pillar version 1.4.4 Index]