extra_cols {pillar}	R Documentation

Retrieve information about columns that didn't fit the available width

Description

Formatting a colonnade object may lead to some columns being omitted due to width restrictions. This method returns a character vector that describes each of the omitted columns.

Usage

extra_cols(x, ...)

## S3 method for class 'pillar_squeezed_colonnade'
extra_cols(x, ..., n = Inf)

Arguments

x
The result of squeeze() on a colonnade object

...
Arguments passed to methods.

n
The number of extra columns to return; the returned vector will always contain as many elements as there are extra columns, but elements beyond n will be NA.

Examples

extra_cols(squeeze(colonnade(list(a = 1:3, b = 4:6), width = 8)))

[Package pillar version 1.6.3 Index]