new_pillar_type {pillar}	R Documentation

Prepare a column type for formatting

Description

Calls type_sum() to format the type. Call format() on the result to render column types.

Usage

new_pillar_type(x, ...)

Arguments

x
A vector for which the type is to be retrieved.

...
These dots are for future extensions and must be empty.

Examples

format(new_pillar_type("a"))
format(new_pillar_type(factor("a")))

[Package pillar version 1.9.0 Index]