squeeze {pillar}	R Documentation

Squeeze a colonnade to a fixed width

Description

The squeeze() function usually doesn't need to be called manually. It returns an object suitable for printing and formatting at a fixed width with additional information about omitted columns, which can be retrieved via extra_cols().

Usage

squeeze(x, width = NULL, ...)

Examples

long_string <- list(paste(letters, collapse = " "))
squeeze(colonnade(long_string), width = 40)
squeeze(colonnade(long_string), width = 20)

[Package pillar version 1.6.3 Index]