tbl_nrow {pillar}	R Documentation

Number of rows in a tbl object

Description

This generic will be called by tbl_format_setup() to determine the number of rows in a tbl object.

Usage

tbl_nrow(x, ...)

Arguments

x
A tbl object.

...
These dots are for future extensions and must be empty.

[Package pillar version 1.10.1 Index]