tbl_sum {pillar}	R Documentation

Provide a succinct summary of an object

Description

tbl_sum() gives a brief textual description of a table-like object, which should include the dimensions and the data source in the first element, and additional information in the other elements (such as grouping for dplyr). The default implementation forwards to obj_sum().

Usage

tbl_sum(x)

Arguments

x
Object to summarise.

Details

This generic will be moved to pillar, and reexported from there as soon as it becomes available.

Value

A named character vector, describing the dimensions in the first element and the data source in the name of the first element.

See Also

type_sum()

[Package pillar version 1.6.3 Index]