type_sum {pillar}	R Documentation

Provide a succinct summary of an object

Description

type_sum() gives a brief summary of object type. Objects that commonly occur in a data frame should return a string with four or less characters.

obj_sum() also includes the size of the object if is_vector_s3() is TRUE.

Usage

type_sum(x)

obj_sum(x)

is_vector_s3(x)

Arguments

x
an object to summarise. Generally only methods of atomic vectors and variants have been implemented.

Examples

obj_sum(1:10)
obj_sum(matrix(1:10))
obj_sum(Sys.Date())
obj_sum(Sys.time())
obj_sum(mean)

[Package pillar version 1.3.0 Index]