Customize your tibble subclass

Gain full control over the appearance of the pillars of your tibble subclass in its body. These methods are intended for implementers of subclasses of the "tbl" class. Users will rarely need them.

ctl_new_pillar(controller, x, width, ..., title = NULL)

ctl_new_compound_pillar(controller, x, width, ..., title = NULL)

Arguments

controller

The object of class "tbl" currently printed.

x

A vector, can also be a data frame, array or matrix in ctl_new_compound_pillar()

width

The available width, can be a vector for multiple tiers

...

These dots are for future extensions and must be empty.

title

The title, derived from the name of the column in the data

Details

ctl_new_pillar() is called to construct pillars for regular (one-dimensional) vectors. The default implementation returns an object constructed with pillar(). Extend this method to tweak pillar components returned from the default implementation. Override this method to completely change the appearance of the pillars.

ctl_new_compound_pillar() is called for compound pillars: columns that are data frames, matrices or arrays. The default implementation returns a compound pillar with suitable formatting for the titles and types of the sub-pillar. Users will only rarely need to override this method if ever.

All components must be of the same height. This restriction may be levied in the future.

Implementations should return NULL if none of the data fits the available width.

Examples

# Create pillar objects
ctl_new_pillar(
  palmerpenguins::penguins,
  palmerpenguins::penguins$species[1:3],
  width = 60
)
#> <pillar>
#> <fct> 
#> Adelie
#> Adelie
#> Adelie
ctl_new_pillar(
  palmerpenguins::penguins,
  palmerpenguins::penguins$bill_length_mm[1:3],
  width = 60
)
#> <pillar>
#>                                                        <dbl>
#>                                                         39.1
#>                                                         39.5
#>                                                         40.3

# Packed data frame
ctl_new_compound_pillar(
  tibble::tibble(),
  palmerpenguins::penguins,
  width = 60
)
#> <compound_pillar[5]>
#> species   island    bill_length_mm bill_depth_mm flipper_length_mm
#> <fct>     <fct>              <dbl>         <dbl>             <int>
#> Adelie    Torgersen           39.1          18.7               181
#> Adelie    Torgersen           39.5          17.4               186
#> Adelie    Torgersen           40.3          18                 195
#> Adelie    Torgersen           NA            NA                  NA
#> Adelie    Torgersen           36.7          19.3               193
#> Adelie    Torgersen           39.3          20.6               190
#> Adelie    Torgersen           38.9          17.8               181
#> Adelie    Torgersen           39.2          19.6               195
#> Adelie    Torgersen           34.1          18.1               193
#> Adelie    Torgersen           42            20.2               190
#> Adelie    Torgersen           37.8          17.1               186
#> Adelie    Torgersen           37.8          17.3               180
#> Adelie    Torgersen           41.1          17.6               182
#> Adelie    Torgersen           38.6          21.2               191
#> Adelie    Torgersen           34.6          21.1               198
#> Adelie    Torgersen           36.6          17.8               185
#> Adelie    Torgersen           38.7          19                 195
#> Adelie    Torgersen           42.5          20.7               197
#> Adelie    Torgersen           34.4          18.4               184
#> Adelie    Torgersen           46            21.5               194
#> Adelie    Biscoe              37.8          18.3               174
#> Adelie    Biscoe              37.7          18.7               180
#> Adelie    Biscoe              35.9          19.2               189
#> Adelie    Biscoe              38.2          18.1               185
#> Adelie    Biscoe              38.8          17.2               180
#> Adelie    Biscoe              35.3          18.9               187
#> Adelie    Biscoe              40.6          18.6               183
#> Adelie    Biscoe              40.5          17.9               187
#> Adelie    Biscoe              37.9          18.6               172
#> Adelie    Biscoe              40.5          18.9               180
#> Adelie    Dream               39.5          16.7               178
#> Adelie    Dream               37.2          18.1               178
#> Adelie    Dream               39.5          17.8               188
#> Adelie    Dream               40.9          18.9               184
#> Adelie    Dream               36.4          17                 195
#> Adelie    Dream               39.2          21.1               196
#> Adelie    Dream               38.8          20                 190
#> Adelie    Dream               42.2          18.5               180
#> Adelie    Dream               37.6          19.3               181
#> Adelie    Dream               39.8          19.1               184
#> Adelie    Dream               36.5          18                 182
#> Adelie    Dream               40.8          18.4               195
#> Adelie    Dream               36            18.5               186
#> Adelie    Dream               44.1          19.7               196
#> Adelie    Dream               37            16.9               185
#> Adelie    Dream               39.6          18.8               190
#> Adelie    Dream               41.1          19                 182
#> Adelie    Dream               37.5          18.9               179
#> Adelie    Dream               36            17.9               190
#> Adelie    Dream               42.3          21.2               191
#> Adelie    Biscoe              39.6          17.7               186
#> Adelie    Biscoe              40.1          18.9               188
#> Adelie    Biscoe              35            17.9               190
#> Adelie    Biscoe              42            19.5               200
#> Adelie    Biscoe              34.5          18.1               187
#> Adelie    Biscoe              41.4          18.6               191
#> Adelie    Biscoe              39            17.5               186
#> Adelie    Biscoe              40.6          18.8               193
#> Adelie    Biscoe              36.5          16.6               181
#> Adelie    Biscoe              37.6          19.1               194
#> Adelie    Biscoe              35.7          16.9               185
#> Adelie    Biscoe              41.3          21.1               195
#> Adelie    Biscoe              37.6          17                 185
#> Adelie    Biscoe              41.1          18.2               192
#> Adelie    Biscoe              36.4          17.1               184
#> Adelie    Biscoe              41.6          18                 192
#> Adelie    Biscoe              35.5          16.2               195
#> Adelie    Biscoe              41.1          19.1               188
#> Adelie    Torgersen           35.9          16.6               190
#> Adelie    Torgersen           41.8          19.4               198
#> Adelie    Torgersen           33.5          19                 190
#> Adelie    Torgersen           39.7          18.4               190
#> Adelie    Torgersen           39.6          17.2               196
#> Adelie    Torgersen           45.8          18.9               197
#> Adelie    Torgersen           35.5          17.5               190
#> Adelie    Torgersen           42.8          18.5               195
#> Adelie    Torgersen           40.9          16.8               191
#> Adelie    Torgersen           37.2          19.4               184
#> Adelie    Torgersen           36.2          16.1               187
#> Adelie    Torgersen           42.1          19.1               195
#> Adelie    Torgersen           34.6          17.2               189
#> Adelie    Torgersen           42.9          17.6               196
#> Adelie    Torgersen           36.7          18.8               187
#> Adelie    Torgersen           35.1          19.4               193
#> Adelie    Dream               37.3          17.8               191
#> Adelie    Dream               41.3          20.3               194
#> Adelie    Dream               36.3          19.5               190
#> Adelie    Dream               36.9          18.6               189
#> Adelie    Dream               38.3          19.2               189
#> Adelie    Dream               38.9          18.8               190
#> Adelie    Dream               35.7          18                 202
#> Adelie    Dream               41.1          18.1               205
#> Adelie    Dream               34            17.1               185
#> Adelie    Dream               39.6          18.1               186
#> Adelie    Dream               36.2          17.3               187
#> Adelie    Dream               40.8          18.9               208
#> Adelie    Dream               38.1          18.6               190
#> Adelie    Dream               40.3          18.5               196
#> Adelie    Dream               33.1          16.1               178
#> Adelie    Dream               43.2          18.5               192
#> Adelie    Biscoe              35            17.9               192
#> Adelie    Biscoe              41            20                 203
#> Adelie    Biscoe              37.7          16                 183
#> Adelie    Biscoe              37.8          20                 190
#> Adelie    Biscoe              37.9          18.6               193
#> Adelie    Biscoe              39.7          18.9               184
#> Adelie    Biscoe              38.6          17.2               199
#> Adelie    Biscoe              38.2          20                 190
#> Adelie    Biscoe              38.1          17                 181
#> Adelie    Biscoe              43.2          19                 197
#> Adelie    Biscoe              38.1          16.5               198
#> Adelie    Biscoe              45.6          20.3               191
#> Adelie    Biscoe              39.7          17.7               193
#> Adelie    Biscoe              42.2          19.5               197
#> Adelie    Biscoe              39.6          20.7               191
#> Adelie    Biscoe              42.7          18.3               196
#> Adelie    Torgersen           38.6          17                 188
#> Adelie    Torgersen           37.3          20.5               199
#> Adelie    Torgersen           35.7          17                 189
#> Adelie    Torgersen           41.1          18.6               189
#> Adelie    Torgersen           36.2          17.2               187
#> Adelie    Torgersen           37.7          19.8               198
#> Adelie    Torgersen           40.2          17                 176
#> Adelie    Torgersen           41.4          18.5               202
#> Adelie    Torgersen           35.2          15.9               186
#> Adelie    Torgersen           40.6          19                 199
#> Adelie    Torgersen           38.8          17.6               191
#> Adelie    Torgersen           41.5          18.3               195
#> Adelie    Torgersen           39            17.1               191
#> Adelie    Torgersen           44.1          18                 210
#> Adelie    Torgersen           38.5          17.9               190
#> Adelie    Torgersen           43.1          19.2               197
#> Adelie    Dream               36.8          18.5               193
#> Adelie    Dream               37.5          18.5               199
#> Adelie    Dream               38.1          17.6               187
#> Adelie    Dream               41.1          17.5               190
#> Adelie    Dream               35.6          17.5               191
#> Adelie    Dream               40.2          20.1               200
#> Adelie    Dream               37            16.5               185
#> Adelie    Dream               39.7          17.9               193
#> Adelie    Dream               40.2          17.1               193
#> Adelie    Dream               40.6          17.2               187
#> Adelie    Dream               32.1          15.5               188
#> Adelie    Dream               40.7          17                 190
#> Adelie    Dream               37.3          16.8               192
#> Adelie    Dream               39            18.7               185
#> Adelie    Dream               39.2          18.6               190
#> Adelie    Dream               36.6          18.4               184
#> Adelie    Dream               36            17.8               195
#> Adelie    Dream               37.8          18.1               193
#> Adelie    Dream               36            17.1               187
#> Adelie    Dream               41.5          18.5               201
#> Gentoo    Biscoe              46.1          13.2               211
#> Gentoo    Biscoe              50            16.3               230
#> Gentoo    Biscoe              48.7          14.1               210
#> Gentoo    Biscoe              50            15.2               218
#> Gentoo    Biscoe              47.6          14.5               215
#> Gentoo    Biscoe              46.5          13.5               210
#> Gentoo    Biscoe              45.4          14.6               211
#> Gentoo    Biscoe              46.7          15.3               219
#> Gentoo    Biscoe              43.3          13.4               209
#> Gentoo    Biscoe              46.8          15.4               215
#> Gentoo    Biscoe              40.9          13.7               214
#> Gentoo    Biscoe              49            16.1               216
#> Gentoo    Biscoe              45.5          13.7               214
#> Gentoo    Biscoe              48.4          14.6               213
#> Gentoo    Biscoe              45.8          14.6               210
#> Gentoo    Biscoe              49.3          15.7               217
#> Gentoo    Biscoe              42            13.5               210
#> Gentoo    Biscoe              49.2          15.2               221
#> Gentoo    Biscoe              46.2          14.5               209
#> Gentoo    Biscoe              48.7          15.1               222
#> Gentoo    Biscoe              50.2          14.3               218
#> Gentoo    Biscoe              45.1          14.5               215
#> Gentoo    Biscoe              46.5          14.5               213
#> Gentoo    Biscoe              46.3          15.8               215
#> Gentoo    Biscoe              42.9          13.1               215
#> Gentoo    Biscoe              46.1          15.1               215
#> Gentoo    Biscoe              44.5          14.3               216
#> Gentoo    Biscoe              47.8          15                 215
#> Gentoo    Biscoe              48.2          14.3               210
#> Gentoo    Biscoe              50            15.3               220
#> Gentoo    Biscoe              47.3          15.3               222
#> Gentoo    Biscoe              42.8          14.2               209
#> Gentoo    Biscoe              45.1          14.5               207
#> Gentoo    Biscoe              59.6          17                 230
#> Gentoo    Biscoe              49.1          14.8               220
#> Gentoo    Biscoe              48.4          16.3               220
#> Gentoo    Biscoe              42.6          13.7               213
#> Gentoo    Biscoe              44.4          17.3               219
#> Gentoo    Biscoe              44            13.6               208
#> Gentoo    Biscoe              48.7          15.7               208
#> Gentoo    Biscoe              42.7          13.7               208
#> Gentoo    Biscoe              49.6          16                 225
#> Gentoo    Biscoe              45.3          13.7               210
#> Gentoo    Biscoe              49.6          15                 216
#> Gentoo    Biscoe              50.5          15.9               222
#> Gentoo    Biscoe              43.6          13.9               217
#> Gentoo    Biscoe              45.5          13.9               210
#> Gentoo    Biscoe              50.5          15.9               225
#> Gentoo    Biscoe              44.9          13.3               213
#> Gentoo    Biscoe              45.2          15.8               215
#> Gentoo    Biscoe              46.6          14.2               210
#> Gentoo    Biscoe              48.5          14.1               220
#> Gentoo    Biscoe              45.1          14.4               210
#> Gentoo    Biscoe              50.1          15                 225
#> Gentoo    Biscoe              46.5          14.4               217
#> Gentoo    Biscoe              45            15.4               220
#> Gentoo    Biscoe              43.8          13.9               208
#> Gentoo    Biscoe              45.5          15                 220
#> Gentoo    Biscoe              43.2          14.5               208
#> Gentoo    Biscoe              50.4          15.3               224
#> Gentoo    Biscoe              45.3          13.8               208
#> Gentoo    Biscoe              46.2          14.9               221
#> Gentoo    Biscoe              45.7          13.9               214
#> Gentoo    Biscoe              54.3          15.7               231
#> Gentoo    Biscoe              45.8          14.2               219
#> Gentoo    Biscoe              49.8          16.8               230
#> Gentoo    Biscoe              46.2          14.4               214
#> Gentoo    Biscoe              49.5          16.2               229
#> Gentoo    Biscoe              43.5          14.2               220
#> Gentoo    Biscoe              50.7          15                 223
#> Gentoo    Biscoe              47.7          15                 216
#> Gentoo    Biscoe              46.4          15.6               221
#> Gentoo    Biscoe              48.2          15.6               221
#> Gentoo    Biscoe              46.5          14.8               217
#> Gentoo    Biscoe              46.4          15                 216
#> Gentoo    Biscoe              48.6          16                 230
#> Gentoo    Biscoe              47.5          14.2               209
#> Gentoo    Biscoe              51.1          16.3               220
#> Gentoo    Biscoe              45.2          13.8               215
#> Gentoo    Biscoe              45.2          16.4               223
#> Gentoo    Biscoe              49.1          14.5               212
#> Gentoo    Biscoe              52.5          15.6               221
#> Gentoo    Biscoe              47.4          14.6               212
#> Gentoo    Biscoe              50            15.9               224
#> Gentoo    Biscoe              44.9          13.8               212
#> Gentoo    Biscoe              50.8          17.3               228
#> Gentoo    Biscoe              43.4          14.4               218
#> Gentoo    Biscoe              51.3          14.2               218
#> Gentoo    Biscoe              47.5          14                 212
#> Gentoo    Biscoe              52.1          17                 230
#> Gentoo    Biscoe              47.5          15                 218
#> Gentoo    Biscoe              52.2          17.1               228
#> Gentoo    Biscoe              45.5          14.5               212
#> Gentoo    Biscoe              49.5          16.1               224
#> Gentoo    Biscoe              44.5          14.7               214
#> Gentoo    Biscoe              50.8          15.7               226
#> Gentoo    Biscoe              49.4          15.8               216
#> Gentoo    Biscoe              46.9          14.6               222
#> Gentoo    Biscoe              48.4          14.4               203
#> Gentoo    Biscoe              51.1          16.5               225
#> Gentoo    Biscoe              48.5          15                 219
#> Gentoo    Biscoe              55.9          17                 228
#> Gentoo    Biscoe              47.2          15.5               215
#> Gentoo    Biscoe              49.1          15                 228
#> Gentoo    Biscoe              47.3          13.8               216
#> Gentoo    Biscoe              46.8          16.1               215
#> Gentoo    Biscoe              41.7          14.7               210
#> Gentoo    Biscoe              53.4          15.8               219
#> Gentoo    Biscoe              43.3          14                 208
#> Gentoo    Biscoe              48.1          15.1               209
#> Gentoo    Biscoe              50.5          15.2               216
#> Gentoo    Biscoe              49.8          15.9               229
#> Gentoo    Biscoe              43.5          15.2               213
#> Gentoo    Biscoe              51.5          16.3               230
#> Gentoo    Biscoe              46.2          14.1               217
#> Gentoo    Biscoe              55.1          16                 230
#> Gentoo    Biscoe              44.5          15.7               217
#> Gentoo    Biscoe              48.8          16.2               222
#> Gentoo    Biscoe              47.2          13.7               214
#> Gentoo    Biscoe              NA            NA                  NA
#> Gentoo    Biscoe              46.8          14.3               215
#> Gentoo    Biscoe              50.4          15.7               222
#> Gentoo    Biscoe              45.2          14.8               212
#> Gentoo    Biscoe              49.9          16.1               213
#> Chinstrap Dream               46.5          17.9               192
#> Chinstrap Dream               50            19.5               196
#> Chinstrap Dream               51.3          19.2               193
#> Chinstrap Dream               45.4          18.7               188
#> Chinstrap Dream               52.7          19.8               197
#> Chinstrap Dream               45.2          17.8               198
#> Chinstrap Dream               46.1          18.2               178
#> Chinstrap Dream               51.3          18.2               197
#> Chinstrap Dream               46            18.9               195
#> Chinstrap Dream               51.3          19.9               198
#> Chinstrap Dream               46.6          17.8               193
#> Chinstrap Dream               51.7          20.3               194
#> Chinstrap Dream               47            17.3               185
#> Chinstrap Dream               52            18.1               201
#> Chinstrap Dream               45.9          17.1               190
#> Chinstrap Dream               50.5          19.6               201
#> Chinstrap Dream               50.3          20                 197
#> Chinstrap Dream               58            17.8               181
#> Chinstrap Dream               46.4          18.6               190
#> Chinstrap Dream               49.2          18.2               195
#> Chinstrap Dream               42.4          17.3               181
#> Chinstrap Dream               48.5          17.5               191
#> Chinstrap Dream               43.2          16.6               187
#> Chinstrap Dream               50.6          19.4               193
#> Chinstrap Dream               46.7          17.9               195
#> Chinstrap Dream               52            19                 197
#> Chinstrap Dream               50.5          18.4               200
#> Chinstrap Dream               49.5          19                 200
#> Chinstrap Dream               46.4          17.8               191
#> Chinstrap Dream               52.8          20                 205
#> Chinstrap Dream               40.9          16.6               187
#> Chinstrap Dream               54.2          20.8               201
#> Chinstrap Dream               42.5          16.7               187
#> Chinstrap Dream               51            18.8               203
#> Chinstrap Dream               49.7          18.6               195
#> Chinstrap Dream               47.5          16.8               199
#> Chinstrap Dream               47.6          18.3               195
#> Chinstrap Dream               52            20.7               210
#> Chinstrap Dream               46.9          16.6               192
#> Chinstrap Dream               53.5          19.9               205
#> Chinstrap Dream               49            19.5               210
#> Chinstrap Dream               46.2          17.5               187
#> Chinstrap Dream               50.9          19.1               196
#> Chinstrap Dream               45.5          17                 196
#> Chinstrap Dream               50.9          17.9               196
#> Chinstrap Dream               50.8          18.5               201
#> Chinstrap Dream               50.1          17.9               190
#> Chinstrap Dream               49            19.6               212
#> Chinstrap Dream               51.5          18.7               187
#> Chinstrap Dream               49.8          17.3               198
#> Chinstrap Dream               48.1          16.4               199
#> Chinstrap Dream               51.4          19                 201
#> Chinstrap Dream               45.7          17.3               193
#> Chinstrap Dream               50.7          19.7               203
#> Chinstrap Dream               42.5          17.3               187
#> Chinstrap Dream               52.2          18.8               197
#> Chinstrap Dream               45.2          16.6               191
#> Chinstrap Dream               49.3          19.9               203
#> Chinstrap Dream               50.2          18.8               202
#> Chinstrap Dream               45.6          19.4               194
#> Chinstrap Dream               51.9          19.5               206
#> Chinstrap Dream               46.8          16.5               189
#> Chinstrap Dream               45.7          17                 195
#> Chinstrap Dream               55.8          19.8               207
#> Chinstrap Dream               43.5          18.1               202
#> Chinstrap Dream               49.6          18.2               193
#> Chinstrap Dream               50.8          19                 210
#> Chinstrap Dream               50.2          18.7               198
#> … and 3 more sub-pillars

# Packed matrix
ctl_new_compound_pillar(tibble::tibble(), matrix(1:6, ncol = 2), width = 60)
#> <compound_pillar[2]>
#>  [,1]  [,2]
#> <int> <int>
#>     1     4
#>     2     5
#>     3     6

# Packed array
ctl_new_compound_pillar(tibble::tibble(), Titanic, width = 60)
#> <pillar>
#>                                              <table[,2,2,2]>
#>                                                          0 …
#>                                                          0 …
#>                                                         35 …
#>                                                          0 …

# Customize output
lines <- function(char = "-") {
  stopifnot(nchar(char) == 1)
  structure(char, class = "lines")
}

format.lines <- function(x, width, ...) {
  paste(rep(x, width), collapse = "")
}

ctl_new_pillar.line_tbl <- function(controller, x, width, ..., title = NULL) {
  out <- NextMethod()
  new_pillar(list(
    title = out$title,
    type = out$type,
    lines = new_pillar_component(list(lines("=")), width = 1),
    data = out$data
  ))
}

vctrs::new_data_frame(
  list(a = 1:3, b = letters[1:3]),
  class = c("line_tbl", "tbl")
)
#> # A data frame: 3 × 2
#>       a b    
#>   <int> <chr>
#> 1     1 a    
#> 2     2 b    
#> 3     3 c