Customize your tibble subclass

Gain full control over the appearance of the pillars of your tibble subclass in its body. These methods are intended for implementers of subclasses of the "tbl" class. Users will rarely need them.

ctl_new_pillar(controller, x, width, ..., title = NULL)

ctl_new_pillar_list(
  controller,
  x,
  width,
  ...,
  title = NULL,
  first_pillar = NULL
)

Arguments

controller

The object of class "tbl" currently printed.

x

A vector, can also be a data frame, array or matrix. in ctl_new_pillar_list().

width

The available width, can be a vector for multiple tiers. If NULL, compute only the first pillar.

...

These dots are for future extensions and must be empty.

title

The title, derived from the name of the column in the data.

first_pillar

Can be passed to this method if the first pillar for a compound pillar (or the pillar itself for a simple pillar) has been computed already.

Details

ctl_new_pillar() is called to construct pillars for regular (one-dimensional) vectors. The default implementation returns an object constructed with pillar(). Extend this method to tweak pillar components returned from the default implementation. Override this method to completely change the appearance of the pillars.

ctl_new_pillar_list() is called to construct a list of pillars. It also works for compound pillars: columns that are data frames, matrices or arrays. This method is also called to initiate the construction of all pillars in the tibble to be printed. If called for a regular one-dimensional vector, it returns a list of length one. In any case, all pillars in the returned list of pillars represent only the first column in case of compound columns. This ensures that only those pillars that are shown are constructed. To print all columns of a packed data frame, ctl_new_pillar_list() eventually calls itself recursively. Users will only rarely need to override this method if ever.

All components must be of the same height. This restriction may be levied in the future.

Implementations should return NULL if none of the data fits the available width.

Examples

# Create pillar objects
ctl_new_pillar(
  palmerpenguins::penguins,
  palmerpenguins::penguins$species[1:3],
  width = 60
)
#> <pillar>
#> <fct> 
#> Adelie
#> Adelie
#> Adelie
ctl_new_pillar(
  palmerpenguins::penguins,
  palmerpenguins::penguins$bill_length_mm[1:3],
  width = 60
)
#> <pillar>
#> <dbl>
#>  39.1
#>  39.5
#>  40.3

# Packed data frame
ctl_new_pillar_list(
  tibble::tibble(),
  palmerpenguins::penguins,
  width = 60
)
#> [[1]]
#> <pillar>
#> species  
#> <fct>    
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Adelie   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Gentoo   
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> Chinstrap
#> 
#> [[2]]
#> <pillar>
#> island   
#> <fct>    
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Torgersen
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Biscoe   
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> Dream    
#> 
#> [[3]]
#> <pillar>
#> bill_length_mm
#>          <dbl>
#>           39.1
#>           39.5
#>           40.3
#>           NA  
#>           36.7
#>           39.3
#>           38.9
#>           39.2
#>           34.1
#>           42  
#>           37.8
#>           37.8
#>           41.1
#>           38.6
#>           34.6
#>           36.6
#>           38.7
#>           42.5
#>           34.4
#>           46  
#>           37.8
#>           37.7
#>           35.9
#>           38.2
#>           38.8
#>           35.3
#>           40.6
#>           40.5
#>           37.9
#>           40.5
#>           39.5
#>           37.2
#>           39.5
#>           40.9
#>           36.4
#>           39.2
#>           38.8
#>           42.2
#>           37.6
#>           39.8
#>           36.5
#>           40.8
#>           36  
#>           44.1
#>           37  
#>           39.6
#>           41.1
#>           37.5
#>           36  
#>           42.3
#>           39.6
#>           40.1
#>           35  
#>           42  
#>           34.5
#>           41.4
#>           39  
#>           40.6
#>           36.5
#>           37.6
#>           35.7
#>           41.3
#>           37.6
#>           41.1
#>           36.4
#>           41.6
#>           35.5
#>           41.1
#>           35.9
#>           41.8
#>           33.5
#>           39.7
#>           39.6
#>           45.8
#>           35.5
#>           42.8
#>           40.9
#>           37.2
#>           36.2
#>           42.1
#>           34.6
#>           42.9
#>           36.7
#>           35.1
#>           37.3
#>           41.3
#>           36.3
#>           36.9
#>           38.3
#>           38.9
#>           35.7
#>           41.1
#>           34  
#>           39.6
#>           36.2
#>           40.8
#>           38.1
#>           40.3
#>           33.1
#>           43.2
#>           35  
#>           41  
#>           37.7
#>           37.8
#>           37.9
#>           39.7
#>           38.6
#>           38.2
#>           38.1
#>           43.2
#>           38.1
#>           45.6
#>           39.7
#>           42.2
#>           39.6
#>           42.7
#>           38.6
#>           37.3
#>           35.7
#>           41.1
#>           36.2
#>           37.7
#>           40.2
#>           41.4
#>           35.2
#>           40.6
#>           38.8
#>           41.5
#>           39  
#>           44.1
#>           38.5
#>           43.1
#>           36.8
#>           37.5
#>           38.1
#>           41.1
#>           35.6
#>           40.2
#>           37  
#>           39.7
#>           40.2
#>           40.6
#>           32.1
#>           40.7
#>           37.3
#>           39  
#>           39.2
#>           36.6
#>           36  
#>           37.8
#>           36  
#>           41.5
#>           46.1
#>           50  
#>           48.7
#>           50  
#>           47.6
#>           46.5
#>           45.4
#>           46.7
#>           43.3
#>           46.8
#>           40.9
#>           49  
#>           45.5
#>           48.4
#>           45.8
#>           49.3
#>           42  
#>           49.2
#>           46.2
#>           48.7
#>           50.2
#>           45.1
#>           46.5
#>           46.3
#>           42.9
#>           46.1
#>           44.5
#>           47.8
#>           48.2
#>           50  
#>           47.3
#>           42.8
#>           45.1
#>           59.6
#>           49.1
#>           48.4
#>           42.6
#>           44.4
#>           44  
#>           48.7
#>           42.7
#>           49.6
#>           45.3
#>           49.6
#>           50.5
#>           43.6
#>           45.5
#>           50.5
#>           44.9
#>           45.2
#>           46.6
#>           48.5
#>           45.1
#>           50.1
#>           46.5
#>           45  
#>           43.8
#>           45.5
#>           43.2
#>           50.4
#>           45.3
#>           46.2
#>           45.7
#>           54.3
#>           45.8
#>           49.8
#>           46.2
#>           49.5
#>           43.5
#>           50.7
#>           47.7
#>           46.4
#>           48.2
#>           46.5
#>           46.4
#>           48.6
#>           47.5
#>           51.1
#>           45.2
#>           45.2
#>           49.1
#>           52.5
#>           47.4
#>           50  
#>           44.9
#>           50.8
#>           43.4
#>           51.3
#>           47.5
#>           52.1
#>           47.5
#>           52.2
#>           45.5
#>           49.5
#>           44.5
#>           50.8
#>           49.4
#>           46.9
#>           48.4
#>           51.1
#>           48.5
#>           55.9
#>           47.2
#>           49.1
#>           47.3
#>           46.8
#>           41.7
#>           53.4
#>           43.3
#>           48.1
#>           50.5
#>           49.8
#>           43.5
#>           51.5
#>           46.2
#>           55.1
#>           44.5
#>           48.8
#>           47.2
#>           NA  
#>           46.8
#>           50.4
#>           45.2
#>           49.9
#>           46.5
#>           50  
#>           51.3
#>           45.4
#>           52.7
#>           45.2
#>           46.1
#>           51.3
#>           46  
#>           51.3
#>           46.6
#>           51.7
#>           47  
#>           52  
#>           45.9
#>           50.5
#>           50.3
#>           58  
#>           46.4
#>           49.2
#>           42.4
#>           48.5
#>           43.2
#>           50.6
#>           46.7
#>           52  
#>           50.5
#>           49.5
#>           46.4
#>           52.8
#>           40.9
#>           54.2
#>           42.5
#>           51  
#>           49.7
#>           47.5
#>           47.6
#>           52  
#>           46.9
#>           53.5
#>           49  
#>           46.2
#>           50.9
#>           45.5
#>           50.9
#>           50.8
#>           50.1
#>           49  
#>           51.5
#>           49.8
#>           48.1
#>           51.4
#>           45.7
#>           50.7
#>           42.5
#>           52.2
#>           45.2
#>           49.3
#>           50.2
#>           45.6
#>           51.9
#>           46.8
#>           45.7
#>           55.8
#>           43.5
#>           49.6
#>           50.8
#>           50.2
#> 
#> [[4]]
#> <pillar>
#> bill_depth_mm
#>         <dbl>
#>          18.7
#>          17.4
#>          18  
#>          NA  
#>          19.3
#>          20.6
#>          17.8
#>          19.6
#>          18.1
#>          20.2
#>          17.1
#>          17.3
#>          17.6
#>          21.2
#>          21.1
#>          17.8
#>          19  
#>          20.7
#>          18.4
#>          21.5
#>          18.3
#>          18.7
#>          19.2
#>          18.1
#>          17.2
#>          18.9
#>          18.6
#>          17.9
#>          18.6
#>          18.9
#>          16.7
#>          18.1
#>          17.8
#>          18.9
#>          17  
#>          21.1
#>          20  
#>          18.5
#>          19.3
#>          19.1
#>          18  
#>          18.4
#>          18.5
#>          19.7
#>          16.9
#>          18.8
#>          19  
#>          18.9
#>          17.9
#>          21.2
#>          17.7
#>          18.9
#>          17.9
#>          19.5
#>          18.1
#>          18.6
#>          17.5
#>          18.8
#>          16.6
#>          19.1
#>          16.9
#>          21.1
#>          17  
#>          18.2
#>          17.1
#>          18  
#>          16.2
#>          19.1
#>          16.6
#>          19.4
#>          19  
#>          18.4
#>          17.2
#>          18.9
#>          17.5
#>          18.5
#>          16.8
#>          19.4
#>          16.1
#>          19.1
#>          17.2
#>          17.6
#>          18.8
#>          19.4
#>          17.8
#>          20.3
#>          19.5
#>          18.6
#>          19.2
#>          18.8
#>          18  
#>          18.1
#>          17.1
#>          18.1
#>          17.3
#>          18.9
#>          18.6
#>          18.5
#>          16.1
#>          18.5
#>          17.9
#>          20  
#>          16  
#>          20  
#>          18.6
#>          18.9
#>          17.2
#>          20  
#>          17  
#>          19  
#>          16.5
#>          20.3
#>          17.7
#>          19.5
#>          20.7
#>          18.3
#>          17  
#>          20.5
#>          17  
#>          18.6
#>          17.2
#>          19.8
#>          17  
#>          18.5
#>          15.9
#>          19  
#>          17.6
#>          18.3
#>          17.1
#>          18  
#>          17.9
#>          19.2
#>          18.5
#>          18.5
#>          17.6
#>          17.5
#>          17.5
#>          20.1
#>          16.5
#>          17.9
#>          17.1
#>          17.2
#>          15.5
#>          17  
#>          16.8
#>          18.7
#>          18.6
#>          18.4
#>          17.8
#>          18.1
#>          17.1
#>          18.5
#>          13.2
#>          16.3
#>          14.1
#>          15.2
#>          14.5
#>          13.5
#>          14.6
#>          15.3
#>          13.4
#>          15.4
#>          13.7
#>          16.1
#>          13.7
#>          14.6
#>          14.6
#>          15.7
#>          13.5
#>          15.2
#>          14.5
#>          15.1
#>          14.3
#>          14.5
#>          14.5
#>          15.8
#>          13.1
#>          15.1
#>          14.3
#>          15  
#>          14.3
#>          15.3
#>          15.3
#>          14.2
#>          14.5
#>          17  
#>          14.8
#>          16.3
#>          13.7
#>          17.3
#>          13.6
#>          15.7
#>          13.7
#>          16  
#>          13.7
#>          15  
#>          15.9
#>          13.9
#>          13.9
#>          15.9
#>          13.3
#>          15.8
#>          14.2
#>          14.1
#>          14.4
#>          15  
#>          14.4
#>          15.4
#>          13.9
#>          15  
#>          14.5
#>          15.3
#>          13.8
#>          14.9
#>          13.9
#>          15.7
#>          14.2
#>          16.8
#>          14.4
#>          16.2
#>          14.2
#>          15  
#>          15  
#>          15.6
#>          15.6
#>          14.8
#>          15  
#>          16  
#>          14.2
#>          16.3
#>          13.8
#>          16.4
#>          14.5
#>          15.6
#>          14.6
#>          15.9
#>          13.8
#>          17.3
#>          14.4
#>          14.2
#>          14  
#>          17  
#>          15  
#>          17.1
#>          14.5
#>          16.1
#>          14.7
#>          15.7
#>          15.8
#>          14.6
#>          14.4
#>          16.5
#>          15  
#>          17  
#>          15.5
#>          15  
#>          13.8
#>          16.1
#>          14.7
#>          15.8
#>          14  
#>          15.1
#>          15.2
#>          15.9
#>          15.2
#>          16.3
#>          14.1
#>          16  
#>          15.7
#>          16.2
#>          13.7
#>          NA  
#>          14.3
#>          15.7
#>          14.8
#>          16.1
#>          17.9
#>          19.5
#>          19.2
#>          18.7
#>          19.8
#>          17.8
#>          18.2
#>          18.2
#>          18.9
#>          19.9
#>          17.8
#>          20.3
#>          17.3
#>          18.1
#>          17.1
#>          19.6
#>          20  
#>          17.8
#>          18.6
#>          18.2
#>          17.3
#>          17.5
#>          16.6
#>          19.4
#>          17.9
#>          19  
#>          18.4
#>          19  
#>          17.8
#>          20  
#>          16.6
#>          20.8
#>          16.7
#>          18.8
#>          18.6
#>          16.8
#>          18.3
#>          20.7
#>          16.6
#>          19.9
#>          19.5
#>          17.5
#>          19.1
#>          17  
#>          17.9
#>          18.5
#>          17.9
#>          19.6
#>          18.7
#>          17.3
#>          16.4
#>          19  
#>          17.3
#>          19.7
#>          17.3
#>          18.8
#>          16.6
#>          19.9
#>          18.8
#>          19.4
#>          19.5
#>          16.5
#>          17  
#>          19.8
#>          18.1
#>          18.2
#>          19  
#>          18.7
#> 
#> [[5]]
#> <pillar>
#> flipper_length_mm
#>             <int>
#>               181
#>               186
#>               195
#>                NA
#>               193
#>               190
#>               181
#>               195
#>               193
#>               190
#>               186
#>               180
#>               182
#>               191
#>               198
#>               185
#>               195
#>               197
#>               184
#>               194
#>               174
#>               180
#>               189
#>               185
#>               180
#>               187
#>               183
#>               187
#>               172
#>               180
#>               178
#>               178
#>               188
#>               184
#>               195
#>               196
#>               190
#>               180
#>               181
#>               184
#>               182
#>               195
#>               186
#>               196
#>               185
#>               190
#>               182
#>               179
#>               190
#>               191
#>               186
#>               188
#>               190
#>               200
#>               187
#>               191
#>               186
#>               193
#>               181
#>               194
#>               185
#>               195
#>               185
#>               192
#>               184
#>               192
#>               195
#>               188
#>               190
#>               198
#>               190
#>               190
#>               196
#>               197
#>               190
#>               195
#>               191
#>               184
#>               187
#>               195
#>               189
#>               196
#>               187
#>               193
#>               191
#>               194
#>               190
#>               189
#>               189
#>               190
#>               202
#>               205
#>               185
#>               186
#>               187
#>               208
#>               190
#>               196
#>               178
#>               192
#>               192
#>               203
#>               183
#>               190
#>               193
#>               184
#>               199
#>               190
#>               181
#>               197
#>               198
#>               191
#>               193
#>               197
#>               191
#>               196
#>               188
#>               199
#>               189
#>               189
#>               187
#>               198
#>               176
#>               202
#>               186
#>               199
#>               191
#>               195
#>               191
#>               210
#>               190
#>               197
#>               193
#>               199
#>               187
#>               190
#>               191
#>               200
#>               185
#>               193
#>               193
#>               187
#>               188
#>               190
#>               192
#>               185
#>               190
#>               184
#>               195
#>               193
#>               187
#>               201
#>               211
#>               230
#>               210
#>               218
#>               215
#>               210
#>               211
#>               219
#>               209
#>               215
#>               214
#>               216
#>               214
#>               213
#>               210
#>               217
#>               210
#>               221
#>               209
#>               222
#>               218
#>               215
#>               213
#>               215
#>               215
#>               215
#>               216
#>               215
#>               210
#>               220
#>               222
#>               209
#>               207
#>               230
#>               220
#>               220
#>               213
#>               219
#>               208
#>               208
#>               208
#>               225
#>               210
#>               216
#>               222
#>               217
#>               210
#>               225
#>               213
#>               215
#>               210
#>               220
#>               210
#>               225
#>               217
#>               220
#>               208
#>               220
#>               208
#>               224
#>               208
#>               221
#>               214
#>               231
#>               219
#>               230
#>               214
#>               229
#>               220
#>               223
#>               216
#>               221
#>               221
#>               217
#>               216
#>               230
#>               209
#>               220
#>               215
#>               223
#>               212
#>               221
#>               212
#>               224
#>               212
#>               228
#>               218
#>               218
#>               212
#>               230
#>               218
#>               228
#>               212
#>               224
#>               214
#>               226
#>               216
#>               222
#>               203
#>               225
#>               219
#>               228
#>               215
#>               228
#>               216
#>               215
#>               210
#>               219
#>               208
#>               209
#>               216
#>               229
#>               213
#>               230
#>               217
#>               230
#>               217
#>               222
#>               214
#>                NA
#>               215
#>               222
#>               212
#>               213
#>               192
#>               196
#>               193
#>               188
#>               197
#>               198
#>               178
#>               197
#>               195
#>               198
#>               193
#>               194
#>               185
#>               201
#>               190
#>               201
#>               197
#>               181
#>               190
#>               195
#>               181
#>               191
#>               187
#>               193
#>               195
#>               197
#>               200
#>               200
#>               191
#>               205
#>               187
#>               201
#>               187
#>               203
#>               195
#>               199
#>               195
#>               210
#>               192
#>               205
#>               210
#>               187
#>               196
#>               196
#>               196
#>               201
#>               190
#>               212
#>               187
#>               198
#>               199
#>               201
#>               193
#>               203
#>               187
#>               197
#>               191
#>               203
#>               202
#>               194
#>               206
#>               189
#>               195
#>               207
#>               202
#>               193
#>               210
#>               198
#> 
#> attr(,"extra")
#> [1] "body_mass_g" "sex"         "year"       
#> attr(,"remaining_width")
#> [1] 0
#> attr(,"simple")
#> [1] FALSE

# Packed matrix
ctl_new_pillar_list(tibble::tibble(), matrix(1:6, ncol = 2), width = 60)
#> [[1]]
#> <pillar>
#>  [,1]
#> <int>
#>     1
#>     2
#>     3
#> 
#> [[2]]
#> <pillar>
#>  [,2]
#> <int>
#>     4
#>     5
#>     6
#> 
#> attr(,"extra")
#> integer(0)
#> attr(,"remaining_width")
#> [1] 48
#> attr(,"simple")
#> [1] FALSE

# Packed array
ctl_new_pillar_list(tibble::tibble(), Titanic, width = 60)
#> [[1]]
#> <pillar>
#> <table[,2,2,2]>
#>             0 …
#>             0 …
#>            35 …
#>             0 …
#> 
#> attr(,"remaining_width")
#> [1] 54
#> attr(,"simple")
#> [1] TRUE

# Customize output
lines <- function(char = "-") {
  stopifnot(nchar(char) == 1)
  structure(char, class = "lines")
}

format.lines <- function(x, width, ...) {
  paste(rep(x, width), collapse = "")
}

ctl_new_pillar.line_tbl <- function(controller, x, width, ..., title = NULL) {
  out <- NextMethod()
  new_pillar(list(
    title = out$title,
    type = out$type,
    lines = new_pillar_component(list(lines("=")), width = 1),
    data = out$data
  ))
}

vctrs::new_data_frame(
  list(a = 1:3, b = letters[1:3]),
  class = c("line_tbl", "tbl")
)
#> # A data frame: 3 × 2
#>       a b    
#>   <int> <chr>
#> 1     1 a    
#> 2     2 b    
#> 3     3 c