format_tbl {pillar}	R Documentation

Formatting of tbl objects

Description

These functions and methods are responsible for printing objects of the "tbl" class, which includes tibbles and dbplyr lazy tables. See tibble::formatting for user level documentation, and vignette("customization") for details.

While it is possible to implement a custom format() or print() method for your tibble-like objects, it should never be necessary if your class inherits from "tbl". In this case, the default methods offer many customization options at every level of detail. This means you only need to override or extend implementations for the parts that need change.

Usage

## S3 method for class 'tbl'
print(x, width = NULL, ..., n = NULL, n_extra = NULL)

## S3 method for class 'tbl'
format(x, width = NULL, ..., n = NULL, n_extra = NULL)

Arguments

x
Object to format or print.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns.

...
Passed on to tbl_format_setup().

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, the value of the tibble.max_extra_cols option is used.

See Also

tbl_format_setup() for preparing an object for formatting

Examples

print(vctrs::new_data_frame(list(a = 1), class = "tbl"))

[Package pillar version 1.5.0 Index]