Format a type summary

Called on values returned from type_sum() for defining the description in the capital. Implement this method for a helper class for custom formatting, and return an object of this helper class from your type_sum() implementation. See examples.

format_type_sum(x, width, ...)

# S3 method for default
format_type_sum(x, width, ...)

Arguments

x
A return value from type_sum()

width
The desired total width. If the returned string still is wider, it will be trimmed. Can be NULL.

...
Arguments passed to methods.

Examples

format_type_sum(1, NULL)

#> [1] "\033[3m\033[38;5;246m<1>\033[39m\033[23m"

pillar(1)

#> <pillar>
#> <dbl>
#>     1

type_sum.accel <- function(x) {
  structure("kg m/s^2", width = 8, class = "type_sum_accel")
}
format_type_sum.type_sum_accel <- function(x, width, ...) {
  if (!is.null(width) && width < attr(x, "width")) {
    x <- substr(x, 1, width)
  }
  style_subtle(x)
}
accel <- structure(9.81, class = "accel")
pillar(accel)

#> <pillar>
#> <accel>
#> 9.81