Helper to define the contents of a pillar

This function is useful if your data renders differently depending on the available width. In this case, implement the pillar_shaft() method for your class to return a subclass of "pillar_shaft" and have the format() method for this subclass call new_ornament(). See the implementation of pillar_shaft.numeric() and format.pillar_shaft_decimal() for an example.

Usage

new_ornament(x, width = NULL, align = NULL)

Arguments

x

A character vector with formatting, can use ANYI styles e.g provided by the cli package.

width

An optional width of the resulting pillar, computed from x if missing

align

Alignment, one of "left" or "right"

Examples

new_ornament(c("abc", "de"), align = "right")
#> <pillar_ornament>
#> abc
#>  de