pillar-package {pillar}	R Documentation

pillar: Coloured Formatting for Columns

Description

Formats tabular data in columns or rows using the full range of colours provided by modern terminals. Provides various generics for making every aspect of the display customizable.

Author(s)

Maintainer: Kirill Müller krlmlr+r@mailbox.org

Authors:

Hadley Wickham

Other contributors:

RStudio [copyright holder]

See Also

pillar() for formatting a single column,

print.tbl() for formatting data-frame-like objects,

pillar_options for a list of package options.

Examples

pillar(1:3)
pillar(c(1, 2, 3))
pillar(factor(letters[1:3]), title = "letters")
colonnade(iris[1:3, ])

[Package pillar version 1.6.3 Index]