Package options

Options that affect display of tibble-like output.

pillar_options

Details

These options can be set via options() and queried via getOption(). For this, add a pillar. prefix (the package name and a dot) to the option name. Example: for an option foo, use options(pillar.foo = value) to set it and getOption("pillar.foo") to retrieve the current value. An option value of NULL means that the default is used.

Options for the pillar package

print_max: Maximum number of rows printed, default: 20. Set to Inf to always print all rows. For compatibility reasons, getOption("tibble.print_max") and getOption("dplyr.print_max") are also consulted, this will be soft-deprecated in pillar v2.0.0.

print_min: Number of rows printed if the table has more than print_max rows, default: 10. For compatibility reasons, getOption("tibble.print_min") and getOption("dplyr.print_min") are also consulted, this will be soft-deprecated in pillar v2.0.0.

width: Output width. Default: NULL (use getOption("width")). This can be larger than getOption("width"), in this case the output of the table's body is distributed over multiple tiers for wide tibbles. For compatibility reasons, getOption("tibble.width") and getOption("dplyr.width") are also consulted, this will be soft-deprecated in pillar v2.0.0.

max_footer_lines: The maximum number of lines in the footer, default: 7. Set to Inf to turn off truncation of footer lines. The max_extra_cols option still limits the number of columns printed.

max_extra_cols: The maximum number of columns printed in the footer, default: 100. Set to Inf to show all columns. Set the more predictable max_footer_lines to control the number of footer lines instead.

bold: Use bold font, e.g. for column headers? This currently defaults to FALSE, because many terminal fonts have poor support for bold fonts.

subtle: Use subtle style, e.g. for row numbers and data types? Default: TRUE.

subtle_num: Use subtle style for insignificant digits? Default: FALSE, is also affected by the subtle option.

neg: Highlight negative numbers? Default: TRUE.

sigfig: The number of significant digits that will be printed and highlighted, default: 3. Set the subtle option to FALSE to turn off highlighting of significant digits.

min_title_chars: The minimum number of characters for the column title, default: 15. Column titles may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of column titles.

min_chars: The minimum number of characters wide to display character columns, default: 3. Character columns may be truncated up to that width to save horizontal space. Set to Inf to turn off truncation of character columns.

max_dec_width: The maximum allowed width for decimal notation, default: 13.

bidi: Set to TRUE for experimental support for bidirectional scripts. Default: FALSE. When this option is set, "left right override" and "first strong isolate" Unicode controls are inserted to ensure that text appears in its intended direction and that the column headings correspond to the correct columns.

Examples

# Default setting:
getOption("pillar.sigfig")
#> NULL
pillar(1.234567)
#> <pillar>
#> <dbl>
#>  1.23

# Change for the duration of the session:
old <- options(pillar.sigfig = 6)
pillar(1.234567)
#> <pillar>
#>   <dbl>
#> 1.23457

# Change back to the original value:
options(old)
pillar(1.234567)
#> <pillar>
#> <dbl>
#>  1.23

# Local scope:
local({
  rlang::local_options(pillar.sigfig = 6)
  pillar(1.234567)
})
#> <pillar>
#>   <dbl>
#> 1.23457
pillar(1.234567)
#> <pillar>
#> <dbl>
#>  1.23