Squeeze a colonnade to a fixed width

The squeeze() function usually doesn't need to be called manually. It returns an object suitable for printing and formatting at a fixed width with additional information about omitted columns, which can be retrieved via extra_cols().

squeeze(x, width = NULL, ...)

Examples

long_string <- list(paste(letters, collapse = " "))
squeeze(colonnade(long_string), width = 40)
#> Warning: `squeeze()` was deprecated in pillar 1.5.0.
#>   <chr>                                 
#> 1 a b c d e f g h i j k l m n o p q r s…
squeeze(colonnade(long_string), width = 20)
#> Warning: `squeeze()` was deprecated in pillar 1.5.0.
#>   <chr>             
#> 1 a b c d e f g h i…