Format the body of a tibble

For easier customization, the formatting of a tibble is split into three components: header, body, and footer. The tbl_format_body() method is responsible for formatting the body of a tibble.

Override this method if you need to change the appearance of all parts of the body. If you only need to change the appearance of a single data type, override vctrs::vec_ptype_abbr() and pillar_shaft() for this data type.

tbl_format_body(x, setup, ...)

Arguments

x

A tibble-like object.

setup

A setup object returned from tbl_format_setup().

...

These dots are for future extensions and must be empty.

Value

A character vector.

Examples

setup <- tbl_format_setup(palmerpenguins::penguins)
tbl_format_body(palmerpenguins::penguins, setup)
#>    species island    bill_length_mm bill_depth_mm flipper_le…¹ bod…² sex    year
#>    <fct>   <fct>              <dbl>         <dbl>        <int> <int> <fct> <int>
#>  1 Adelie  Torgersen           39.1          18.7          181  3750 male   2007
#>  2 Adelie  Torgersen           39.5          17.4          186  3800 fema…  2007
#>  3 Adelie  Torgersen           40.3          18            195  3250 fema…  2007
#>  4 Adelie  Torgersen           NA            NA             NA    NA NA     2007
#>  5 Adelie  Torgersen           36.7          19.3          193  3450 fema…  2007
#>  6 Adelie  Torgersen           39.3          20.6          190  3650 male   2007
#>  7 Adelie  Torgersen           38.9          17.8          181  3625 fema…  2007
#>  8 Adelie  Torgersen           39.2          19.6          195  4675 male   2007
#>  9 Adelie  Torgersen           34.1          18.1          193  3475 NA     2007
#> 10 Adelie  Torgersen           42            20.2          190  4250 NA     2007

# Shortcut for debugging
tbl_format_body(setup)
#> <tbl_format_body(setup)>
#>    species island    bill_length_mm bill_depth_mm flipper_le…¹ bod…² sex    year
#>    <fct>   <fct>              <dbl>         <dbl>        <int> <int> <fct> <int>
#>  1 Adelie  Torgersen           39.1          18.7          181  3750 male   2007
#>  2 Adelie  Torgersen           39.5          17.4          186  3800 fema…  2007
#>  3 Adelie  Torgersen           40.3          18            195  3250 fema…  2007
#>  4 Adelie  Torgersen           NA            NA             NA    NA NA     2007
#>  5 Adelie  Torgersen           36.7          19.3          193  3450 fema…  2007
#>  6 Adelie  Torgersen           39.3          20.6          190  3650 male   2007
#>  7 Adelie  Torgersen           38.9          17.8          181  3625 fema…  2007
#>  8 Adelie  Torgersen           39.2          19.6          195  4675 male   2007
#>  9 Adelie  Torgersen           34.1          18.1          193  3475 NA     2007
#> 10 Adelie  Torgersen           42            20.2          190  4250 NA     2007