Format the body of a tibble

For easier customization, the formatting of a tibble is split into three components: header, body, and footer. The tbl_format_body() method is responsible for formatting the body of a tibble.

Override this method if you need to change the appearance of all parts of the body. If you only need to change the appearance of a single data type, override vctrs::vec_ptype_abbr() and pillar_shaft() for this data type.

tbl_format_body(x, setup, ...)

Arguments

x
A tibble-like object.

setup
A setup object returned from tbl_format_setup().

...
These dots are for future extensions and must be empty.

Value

A character vector.