type_sum {pillar}	R Documentation

Provide a succinct summary of an object

Description

type_sum() gives a brief summary of object type. Objects that commonly occur in a data frame should return a string with four or less characters. For most inputs, the argument is forwarded to vctrs::vec_ptype_abbr().

obj_sum() also includes the size of the object if vctrs::vec_is() is TRUE.

Usage

type_sum(x)

obj_sum(x)

Arguments

x
an object to summarise. Generally only methods of atomic vectors and variants have been implemented.

Lifecycle

type_sum() is in the "questioning" stage, because vctrs::vec_ptype_abbr() provides essentially the same functionality based on a robust framework. The generic will remain available for a while. Users and implementers are encouraged to consider vctrs::vec_ptype_abbr() once it becomes stable.

Examples

obj_sum(1:10)
obj_sum(matrix(1:10))
obj_sum(Sys.Date())
obj_sum(Sys.time())
obj_sum(mean)

[Package pillar version 1.4.4 Index]