Provide a succinct summary of an object

type_sum() gives a brief summary of object type. Objects that commonly occur in a data frame should return a string with four or less characters. For most inputs, the argument is forwarded to vctrs::vec_ptype_abbr().

obj_sum() also includes the size of the object if vctrs::vec_is() is TRUE. It should always return a string (a character vector of length one).

size_sum() is called by obj_sum() to format the size of the object. It should always return a string (a character vector of length one), it can be an empty string "" to omit size information, this is what the default method does for scalars.

type_sum(x)

obj_sum(x)

size_sum(x)

Arguments

x
an object to summarise. Generally only methods of atomic vectors and variants have been implemented.

Lifecycle

type_sum() is in the "questioning" stage, because vctrs::vec_ptype_abbr() provides essentially the same functionality based on a robust framework. The generic will remain available for a while. Users and implementers are encouraged to consider vctrs::vec_ptype_abbr() once it becomes stable.

Examples

obj_sum(1:10)

#> [1] "int [10]"

obj_sum(matrix(1:10))

#> [1] "int[,1] [10 × 1]"

obj_sum(Sys.Date())

#> [1] "date [1]"

obj_sum(Sys.time())

#> [1] "dttm [1]"

obj_sum(mean)

#> [1] "fn"

size_sum(1:10)

#> [1] "[10]"

size_sum(trees)

#> [1] "[31 × 3]"

size_sum(Titanic)

#> [1] "[4 × 2 × 2 × 2]"